% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.R
\name{plotpal}
\alias{plotpal}
\title{plotpal
creates a simple plot to enable a quick look at available palettes and corresponding hex codes. Palettes include: spec, deepspec, lightspec, lighterspec, pinks, pinks2, reds, yellows, greens, greens2, blues, blues2, mono, tonka, bellsprout, dockers, fruitypebbles, day, daylong, and army. Default palette is "deepspec".}
\usage{
plotpal(palette = "deepspec")
}
\arguments{
\item{palette}{name of palette to be used in string format}
}
\value{

}
\description{
plotpal
creates a simple plot to enable a quick look at available palettes and corresponding hex codes. Palettes include: spec, deepspec, lightspec, lighterspec, pinks, pinks2, reds, yellows, greens, greens2, blues, blues2, mono, tonka, bellsprout, dockers, fruitypebbles, day, daylong, and army. Default palette is "deepspec".
}
\examples{
plotpal(palette = "day")
plotpal(palette = "daylong")
}
