% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.R
\name{scale_color_fb}
\alias{scale_color_fb}
\title{scale_color_fb
returns a discrete or continuous palette for ggplot2 color. Palettes include: spec, deepspec, lightspec, lighterspec, pinks, pinks2, reds, yellows, greens, greens2, blues, blues2, mono, tonka, bellsprout, dockers, fruitypebbles, day, daylong, and army. Default palette is "deepspec".}
\usage{
scale_color_fb(palette = "deepspec", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{name of palette to be used in string format}

\item{discrete}{set to TRUE if discrete scale desired; FALSE for continuous}

\item{reverse}{default is FALSE; set to TRUE to reverse the order of the palette}

\item{...}{}
}
\value{
returns a ggplot2-compatible color scale based on the template palette given
}
\description{
scale_color_fb
returns a discrete or continuous palette for ggplot2 color. Palettes include: spec, deepspec, lightspec, lighterspec, pinks, pinks2, reds, yellows, greens, greens2, blues, blues2, mono, tonka, bellsprout, dockers, fruitypebbles, day, daylong, and army. Default palette is "deepspec".
}
\examples{
scale_color_fb(palette = "daylong", discrete = TRUE, reverse = FALSE)
}
