% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.R
\name{scale_fill_fb}
\alias{scale_fill_fb}
\title{scale_fill_fb
returns a discrete or continuous palette for ggplot2 fill. Palettes include spec, deepspec, lightspec, lighterspec, pinks, pinks2, reds, yellows, greens, greens2, blues, blues2, mono, tonka, bellsprout, dockers, fruitypebbles, day, daylong, and army. Default palette is "deepspec".}
\usage{
scale_fill_fb(palette = "deepspec", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{name of palette to be used in string format}

\item{discrete}{set to TRUE if discrete scale desired; FALSE for continuous}

\item{reverse}{default is FALSE; set to TRUE to reverse the order of the palette}

\item{...}{}
}
\value{
returns a ggplot2-compatible fill scale based on the template palette given
}
\description{
scale_fill_fb
returns a discrete or continuous palette for ggplot2 fill. Palettes include spec, deepspec, lightspec, lighterspec, pinks, pinks2, reds, yellows, greens, greens2, blues, blues2, mono, tonka, bellsprout, dockers, fruitypebbles, day, daylong, and army. Default palette is "deepspec".
}
\examples{
scale_fill_fb(palette = "deepspec", discrete = TRUE, reverse = FALSE)

# See how the palette ramps for the required number of colors
# the original "deepspec" palette
x <-  letters[1:10]
y <-  rep(1, 10)
df <- data.frame(x, y)
ggplot(df, aes(x=x, y=y, fill=x)) +
    geom_bar(stat="identity") +
    scale_fill_fb(palette = "deepspec", guide = "none") +
    theme(axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          panel.background = element_blank())

# now deepspec with twice the original colors
x <- letters[1:20]
y <- rep(1, 20)
df <- data.frame(x, y)
ggplot(df, (aes(x=x, y=y, fill=x))) +
    geom_bar(stat="identity") +
    scale_fill_fb(palette = "deepspec", guide = "none") +
    theme(axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          panel.background = element_blank())
}
